pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-agent"
version = "0.1.0"
description = "A lean local-LLM agent with persistent workspace and pluggable tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21"]

[project.scripts]
aria          = "aria.main:main"
aria-telegram = "aria.telegram_bot:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"aria" = ["py.typed"]
